+++ crates/core/action-pipeline/src/actions/mod.rs
@@ -1,3 +1,4 @@
 pub mod install_deps;
+pub mod run_target;
 pub mod setup_tool;
 pub mod sync_project;
