+++ crates/core/action-pipeline/src/lib.rs
@@ -2,6 +2,7 @@ mod actions;
 mod errors;
 mod pipeline;
 mod processor;
+mod run_report;
 
 pub use errors::*;
 pub use moon_action_context::*;
