+++ crates/core/action-pipeline/src/run_report.rs
@@ -1,5 +1,5 @@
 use moon_action::Action;
-use moon_runner_context::RunnerContext;
+use moon_action_context::ActionContext;
 use serde::Serialize;
 use std::time::Duration;
 
@@ -8,7 +8,7 @@ use std::time::Duration;
 pub struct RunReport<'a> {
     pub actions: &'a Vec<Action>,
 
-    pub context: &'a RunnerContext,
+    pub context: &'a ActionContext,
 
     /// How long the runner took to execute all actions.
     pub duration: Duration,
@@ -21,7 +21,7 @@ pub struct RunReport<'a> {
 }
 
 impl<'a> RunReport<'a> {
-    pub fn new(actions: &'a Vec<Action>, context: &'a RunnerContext, duration: Duration) -> Self {
+    pub fn new(actions: &'a Vec<Action>, context: &'a ActionContext, duration: Duration) -> Self {
         let mut projected_duration = Duration::new(0, 0);
 
         for action in actions {
