--- crates/core/runner/src/actions/mod.rs
@@ -1,3 +1,0 @@
-mod run_target;
-
-pub use run_target::run_target;
